Death state and way to restart from death.
Added a way to restart upon dying and death state. The enemy controller scipts checks if the enemy and player are close enough. Then in the gamemanager script a boolean 'playerIsDead' is flipped. This bool is needed to be allowed to restart. Restart is then possible by pressen R. When pressing R the tutorial will begin again and the player and enemy are reset to their orinigal position